leetcode/783.二叉搜索树节点最小距离/minDiffInBST_test.go
package _83_二叉搜索树节点最小距离

import "testing"

func Test_minDiffInBST(t *testing.T) {
	type args struct {
		root *TreeNode
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{name: `输入：root = [4,2,6,1,3] 输出：1`, args: args{root: &TreeNode{
			Val: 4,
			Left: &TreeNode{
				Val: 2,
				Left: &TreeNode{
					Val:   1,
					Left:  nil,
					Right: nil,
				},
				Right: &TreeNode{
					Val:   3,
					Left:  nil,
					Right: nil,
				},
			}, Right: &TreeNode{
				Val:   6,
				Left:  nil,
				Right: nil,
			}}}, want: 1},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := minDiffInBST1(tt.args.root); got != tt.want {
				t.Errorf("minDiffInBST1() = %v, want %v", got, tt.want)
			}
			if got := minDiffInBST(tt.args.root); got != tt.want {
				t.Errorf("minDiffInBST() = %v, want %v", got, tt.want)
			}
		})
	}
}
